Moving files that are not exposed in the public API to a new lib folder.
