+++ src/lib/discovery.ts
